+++ Engine/egParticleRenderer.cpp
@@ -53,18 +53,18 @@ namespace Engine::Components
         GetRenderPipeline().SetParam(m_max_scale_size_, scaling_max_slot);
       }
 
-      m_sb_buffer_.SetData(m_instances_.size(), m_instances_.data());
+      m_sb_buffer_.SetData(static_cast<UINT>(m_instances_.size()), m_instances_.data());
       m_sb_buffer_.BindUAV();
 
       const auto thread      = m_cs_->GetThread();
       const auto flatten     = thread[0] * thread[1] * thread[2];
-      const UINT group_count = m_instances_.size() / flatten;
-      const UINT remainder   = m_instances_.size() % flatten;
+      const UINT group_count = static_cast<UINT>(m_instances_.size() / flatten);
+      const UINT remainder   = static_cast<UINT>(m_instances_.size() % flatten);
 
       m_cs_->SetGroup({group_count + (remainder ? 1 : 0), 1, 1});
       m_cs_->Dispatch();
       m_sb_buffer_.UnbindUAV();
-      m_sb_buffer_.GetData(m_instances_.size(), m_instances_.data());
+      m_sb_buffer_.GetData(static_cast<UINT>(m_instances_.size()), m_instances_.data());
     }
 
     // Remove inactive particles.
